+++ github-watcher/src/main/scala/dev/nhyne/template/kustomize/Transformer.scala
@@ -1,0 +1,9 @@
+package dev.nhyne.template.kustomize
+
+object Transformer {
+  /*
+   *  Want to do Kustomize transformer things in here but not sure which ones to pickup yet
+   *
+   */
+
+}
