--- github-watcher/src/main/scala/nhyne/template/kustomize/Transformer.scala
@@ -1,9 +1,0 @@
-package nhyne.template.kustomize
-
-object Transformer {
-  /*
-   *  Want to do Kustomize transformer things in here but not sure which ones to pickup yet
-   *
-   */
-
-}
